+++ zerver/tests/test_message_edit.py
@@ -811,6 +811,7 @@ class EditMessageTest(ZulipTestCase):
                 propagate_mode="change_later",
                 send_notification_to_old_thread=False,
                 send_notification_to_new_thread=False,
+                add_messages_to_history=False,
                 content=None,
                 rendered_content=None,
                 prior_mention_user_ids=set(),
@@ -1745,11 +1746,16 @@ class EditMessageTest(ZulipTestCase):
         from_invite_only: bool,
         history_public_to_subscribers: bool,
         user_messages_created: bool,
+        add_messages_to_history: bool = False,
         to_invite_only: bool = True,
     ) -> None:
         admin_user = self.example_user("iago")
         user_losing_access = self.example_user("cordelia")
-        user_gaining_access = self.example_user("hamlet")
+
+        # Not subscribed to old stream
+        user_gaining_access_1 = self.example_user("hamlet")
+        # Subscribed to both streams after msg was sent in old stream
+        user_gaining_access_2 = self.example_user("aaron")
 
         self.login("iago")
         old_stream = self.make_stream("test move stream", invite_only=from_invite_only)
@@ -1763,12 +1769,16 @@ class EditMessageTest(ZulipTestCase):
         self.subscribe(user_losing_access, old_stream.name)
 
         self.subscribe(admin_user, new_stream.name)
-        self.subscribe(user_gaining_access, new_stream.name)
+        self.subscribe(user_gaining_access_1, new_stream.name)
 
         msg_id = self.send_stream_message(
             admin_user, old_stream.name, topic_name="test", content="First"
         )
         self.send_stream_message(admin_user, old_stream.name, topic_name="test", content="Second")
+
+        # Subscribe user after sending msg
+        self.subscribe(user_gaining_access_2, old_stream.name)
+        self.subscribe(user_gaining_access_2, new_stream.name)
 
         self.assertEqual(
             UserMessage.objects.filter(
@@ -1779,7 +1789,14 @@ class EditMessageTest(ZulipTestCase):
         )
         self.assertEqual(
             UserMessage.objects.filter(
-                user_profile_id=user_gaining_access.id,
+                user_profile_id=user_gaining_access_1.id,
+                message_id=msg_id,
+            ).count(),
+            0,
+        )
+        self.assertEqual(
+            UserMessage.objects.filter(
+                user_profile_id=user_gaining_access_2.id,
                 message_id=msg_id,
             ).count(),
             0,
@@ -1791,6 +1808,7 @@ class EditMessageTest(ZulipTestCase):
                 "message_id": msg_id,
                 "stream_id": new_stream.id,
                 "propagate_mode": "change_all",
+                "add_messages_to_history": orjson.dumps(add_messages_to_history).decode(),
             },
         )
         self.assert_json_success(result)
@@ -1816,17 +1834,37 @@ class EditMessageTest(ZulipTestCase):
         # can see the message.
         self.assertEqual(
             UserMessage.objects.filter(
-                user_profile_id=user_gaining_access.id,
+                user_profile_id=user_gaining_access_1.id,
+                message_id=msg_id,
+            ).count(),
+            1 if user_messages_created else 0,
+        )
+        self.assertEqual(
+            UserMessage.objects.filter(
+                user_profile_id=user_gaining_access_2.id,
                 message_id=msg_id,
             ).count(),
             1 if user_messages_created else 0,
         )
 
-    def test_move_message_from_public_to_private_stream_not_shared_history(self) -> None:
+    def test_move_message_from_public_to_private_stream_not_shared_history_not_add_messages_to_history(
+        self,
+    ) -> None:
+        self.parameterized_test_move_message_involving_private_stream(
+            from_invite_only=False,
+            history_public_to_subscribers=False,
+            user_messages_created=False,
+            add_messages_to_history=False,
+        )
+
+    def test_move_message_from_public_to_private_stream_not_shared_history_add_messages_to_history(
+        self,
+    ) -> None:
         self.parameterized_test_move_message_involving_private_stream(
             from_invite_only=False,
             history_public_to_subscribers=False,
             user_messages_created=True,
+            add_messages_to_history=True,
         )
 
     def test_move_message_from_public_to_private_stream_shared_history(self) -> None:
@@ -1840,6 +1878,7 @@ class EditMessageTest(ZulipTestCase):
         self.parameterized_test_move_message_involving_private_stream(
             from_invite_only=True,
             history_public_to_subscribers=False,
+            add_messages_to_history=True,
             user_messages_created=True,
         )